First step to import of Supporter donations.
Ultimately, I decided I had little choice but to create a database of
supporters.  There is really no easy way to track all this properly with
ledger, since ledger is transaction based and this is data that spans
many transactions.

Indeed, even in the NPO Accounting project, we knew there would be
separate data stores for contacts, etc.  This is an ad-hoc start to
that, I guess.