bug in tests found with transaction ref counting
Now that we are properly nesting transactions, this test is shown to be
incorrect.  Namely, it is indeed correct that the entire add should fail
when we've given something invalid: we don't want to add the requestType
if it wasn't already there if the parameter to add
addRequestConfigurations was invalid.