Rework the reference counter for begin_work/commit
This code was clearly not correct as it was.  The main problem is that I
was envisioning begin_work/commit pairs as always matched, but of course
if we die in the middle of a transaction, the counter has to be reset.
That's why we have to set the $dbh->{HandleError} in new() to a
subroutine that does that.

Also, we need to similarly wrap rollback() calls.  When rolling back, we
have to reset the counter as we're not going to commit() (that's the
point).