Correct test that now fails after recent changes.
This test now started failing after the other corrections made in the
last few commits.

As it turns out, this test was buggy.  We wanted to test a request_type
that didn't exist, and verify that calls to addRequestConfigurations
with bad arguments didn't create the request_type if it didn't already
existed.  That's what the test now does.