Do not attempt to intuit renewal response here.
This code worked reasonable well when there was only one type of renewal
notice in play for a long period of time, but the point here is that we
can have many of them and this script should handle sending out the
different ones at different types.

We probably need an entirely separate script to intuit response.