holdRequest():testing hold on already held request
It's clear from the Supporters.pm code that when an hold is attempted on
an already held request, it simply returns the id of the existing hold.

I don't actually remember what behavior I really wanted here.  There are
certainly possibility for semantic confusion in the API with the current
functionality, since the API caller must actually check to verify if the
hold they got is a new one or just an existing one.

Perhaps this should be rethought.  I left a comment in the test code for
that reason.